extract graph into a separate view
